--- a/.gitignore
+++ b/.gitignore
@@ -9,3 +9,4 @@
 /tmp/
 *.gem
 /lib/*.bundle
+/bin/cell_console