--- a/cellular_automata.gemspec
+++ b/cellular_automata.gemspec
@@ -20,9 +20,12 @@ Gem::Specification.new do |spec|
 
   spec.license       = 'MIT'
 
-  spec.add_runtime_dependency "gosu"
-  spec.add_runtime_dependency "rmagick"
+  spec.extensions << 'ext/cellular_c/extconf.rb'
+
+  spec.add_runtime_dependency "gosu", '~> 0', '>= 0.9'
+  spec.add_runtime_dependency "rmagick", '~> 2', '>= 2.15'
   spec.add_development_dependency "bundler", "~> 1", ">= 1.8"
+  spec.add_development_dependency "rake-compiler", '~> 0', '>= 0.9'
   spec.add_development_dependency "rake", "~> 10.0"
   spec.add_development_dependency 'pry', '~> 0.10'
   spec.add_development_dependency 'byebug', '~> 4'