--- a/ext/cellular_c/cellular_c.c
+++ b/ext/cellular_c/cellular_c.c
@@ -4,15 +4,33 @@
 VALUE CellularC = Qnil;
 void Init_cellular_c();
 VALUE method_neighbor_population_of(VALUE, VALUE, VALUE, VALUE);
+VALUE method_dup_state(VALUE, VALUE);
+VALUE method_tick(VALUE, VALUE);
 
 void Init_cellular_c() {
   CellularC = rb_define_module("CellularC");
   rb_define_singleton_method(CellularC, "neighbor_population_of", method_neighbor_population_of, 3);
+  rb_define_singleton_method(CellularC, "dup_state", method_dup_state, 1);
+}
+
+VALUE method_dup_state(VALUE self, VALUE state) {
+  int height = (int)RARRAY_LEN(state);
+  int width  = (int)RARRAY_LEN(rb_ary_entry(state, 0));
+  VALUE dup =  rb_ary_new2(height);
+  for(int y = 0; y < height; y++) {
+    VALUE row = rb_ary_new2(width);
+    for(int x = 0; x < width; x++) {
+      VALUE val = rb_ary_entry(row, x);
+      rb_ary_push(row, val);
+    }
+    rb_ary_push(dup, row);
+  }
+  return dup;
 }
 
 VALUE method_neighbor_population_of(VALUE self, VALUE state, VALUE rx, VALUE ry) {
-  int height = RARRAY_LEN(state);
-  int width  = RARRAY_LEN(rb_ary_entry(state, 0));
+  int height = (int)RARRAY_LEN(state);
+  int width  = (int)RARRAY_LEN(rb_ary_entry(state, 0));
   int x = FIX2INT(rx);
   int y = FIX2INT(ry);
   int min_x = x > 1 ? x - 1 : 0;