--- a/ext/cellular_c/extconf.rb
+++ b/ext/cellular_c/extconf.rb
@@ -1,0 +1,6 @@
+# ext/extconf.rb
+require 'mkmf'
+extension_name = 'cellular_c'
+dir_config(extension_name)
+create_header
+create_makefile(extension_name)