--- a/lib/cellular_automata.rb
+++ b/lib/cellular_automata.rb
@@ -9,6 +9,7 @@ require "cellular_automata/rule"
 require "cellular_automata/board"
 require "cellular_automata/gui_window"
 require "cellular_automata/gif_writer"
+require 'cellular_c.bundle'
 
 module CellularAutomata
 end