--- a/lib/cellular_automata/version.rb
+++ b/lib/cellular_automata/version.rb
@@ -1,3 +1,3 @@
 module CellularAutomata
-  VERSION = "0.1.3"
+  VERSION = "0.1.4"
 end