Benchmarks:
---
Without C extension:
```
  Rehearsal ------------------------------------------
report   6.740000   0.020000   6.760000 (  6.761006)
--------------------------------- total: 6.760000sec

             user     system      total        real
report   6.680000   0.020000   6.700000 (  6.698803)
```
With C extension for population:
```
Rehearsal ------------------------------------------
report   1.910000   0.010000   1.920000 (  1.930078)
--------------------------------- total: 1.920000sec

             user     system      total        real
report   1.920000   0.010000   1.930000 (  1.930797)
```
Add C for duping array:
```
Rehearsal ------------------------------------------
report   1.510000   0.010000   1.520000 (  1.524112)
--------------------------------- total: 1.520000sec

             user     system      total        real
report   1.540000   0.010000   1.550000 (  1.552010)
```
Use rb_ary_dup:
```
Rehearsal ------------------------------------------
report   1.460000   0.010000   1.470000 (  1.473329)
--------------------------------- total: 1.470000sec

             user     system      total        real
report   1.480000   0.020000   1.500000 (  1.496190)
```
Use C for next_state
```
Please report a bug if this causes problems.
Rehearsal ------------------------------------------
report   1.210000   0.020000   1.230000 (  1.219810)
--------------------------------- total: 1.230000sec

             user     system      total        real
report   1.200000   0.020000   1.220000 (  1.222902)
```

Fix up rule.rb
```
Rehearsal ------------------------------------------
report   1.140000   0.010000   1.150000 (  1.148026)
--------------------------------- total: 1.150000sec

             user     system      total        real
report   1.130000   0.020000   1.150000 (  1.143255)
```